requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
redis>=5.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
crewai>=0.28.0
groq>=0.9.0
chromadb>=0.5.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
